fix: move assertAuthorized to middleware and remove from authentication
